refactor: move formatSchoolName
