+++ src/components/SchoolPicker/formatSchoolName.test.tsx
