+++ src/components/SchoolPicker/formatSchoolName.tsx
