--- src/components/formatSchoolName/index.tsx
@@ -1,1 +1,0 @@
-export { formatSchoolName } from "./formatSchoolName";
